Add skelletal blogging engine experiment
In 2015 I was very frustrated with the state of flat file content management systems, hence I set out to build my own. I never finished it, but I do view it as a nice introduction to the twig template engine for PHP. Unfortunately the code in this commit is all I've ever published with regards to the blogging engine. I still wanted to preserve it as it was my first attempt (that was a bit more serious) of actually publishing something open source and working with git.

* Add assignments to .gitignore
* Add skelletal blogging engine
* Create experiments folder